Enhance Navbar and SidebarMenu for SAO role support. Add conditional rendering for SAO-specific navigation items and update role-based visibility in the sidebar. Refactor routing to include SAO role in protected routes and adjust login navigation accordingly. Remove unused EnrollmentLayout component.
