+++ src/pages/institutionalOfficer/sao/indix.tsx
@@ -1,0 +1,3 @@
+export const SaoOfficer = () => {
+  return <div>Sao</div>;
+};
